nice clean up custom hook
